backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.0
python-dotenv==1.0.0
dnspython==2.4.2
bcrypt==4.1.2
PyJWT==2.8.0
